pyServer/requirements.txt
flask
flask_cors
easyocr
Pillow
numpy
opencv-python